Functions for writing in card have been done.
All the functions for writing from the form to the card have been done
but they aren't fully calidated yet. The validations are only about the
size of the input or the characters used (and they are "verified before",
it means that the user can't write whatever he wants but the characters he
is able to write on). By te way, the user can still send the form with
all the input empty.